requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
boto3==1.34.34
python-dotenv==1.0.1
pydantic==2.6.1 